feat: move icon create new thread into top panel (#3346)
* feat: move icon create new thread into top panel

* chore: update conditional chcck length
